Much progress in getting everything to build and run using Visual Studio 2017.

+++ NetRumble/ScreenManager/LoadingScreen.cs
@@ -132,7 +132,7 @@ namespace NetRumble
                 Vector2 textSize = ScreenManager.Font.MeasureString(message);
                 Vector2 textPosition = (viewportSize - textSize) / 2;
 
-                Color color = new Color(255, 255, 255, TransitionAlpha);
+                Color color = new Color((byte)255, (byte)255, (byte)255, TransitionAlpha);
 
                 // Draw the text.
                 ScreenManager.SpriteBatch.Begin();
